+++ spatial_tk/commands/__init__.py
@@ -1,0 +1,4 @@
+"""
+Command modules for spatial_tk CLI.
+"""
+