+++ xenium_process/utils/__init__.py
@@ -1,4 +1,0 @@
-"""
-Utility functions for xenium_process.
-"""
-